[vfs] move Vfs from core to its own directory
