+++ examples/graphics/demo_coords.cpp
@@ -11,7 +11,7 @@
 #include <xci/graphics/Sprites.h>
 #include <xci/graphics/shape/Rectangle.h>
 #include <xci/graphics/Color.h>
-#include <xci/core/Vfs.h>
+#include <xci/vfs/Vfs.h>
 #include <xci/config.h>
 
 #include <cstdlib>
