+++ examples/widgets/demo_form.cpp
@@ -1,7 +1,7 @@
 // demo_form.cpp created on 2018-06-23 as part of xcikit project
 // https://github.com/rbrich/xcikit
 //
-// Copyright 2018–2022 Radek Brich
+// Copyright 2018–2023 Radek Brich
 // Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #include "MousePosInfo.h"
@@ -11,7 +11,7 @@
 #include <xci/widgets/FpsDisplay.h>
 #include <xci/widgets/Form.h>
 #include <xci/widgets/Label.h>
-#include <xci/core/Vfs.h>
+#include <xci/vfs/Vfs.h>
 #include <xci/config.h>
 
 #include <fmt/core.h>
