+++ examples/widgets/demo_term_widget.cpp
@@ -1,13 +1,13 @@
 // demo_term_widget.cpp created on 2018-07-19 as part of xcikit project
 // https://github.com/rbrich/xcikit
 //
-// Copyright 2018–2022 Radek Brich
+// Copyright 2018–2023 Radek Brich
 // Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #include "graphics/common.h"
 
 #include <xci/widgets/TextTerminal.h>
-#include <xci/core/Vfs.h>
+#include <xci/vfs/Vfs.h>
 #include <xci/core/file.h>
 #include <xci/core/log.h>
 #include <xci/config.h>
