+++ examples/widgets/demo_widgets.cpp
@@ -1,7 +1,7 @@
 // demo_widget.cpp created on 2018-03-20 as part of xcikit project
 // https://github.com/rbrich/xcikit
 //
-// Copyright 2018–2022 Radek Brich
+// Copyright 2018–2023 Radek Brich
 // Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #include "graphics/common.h"
@@ -9,7 +9,7 @@
 #include <xci/widgets/Theme.h>
 #include <xci/widgets/Button.h>
 #include <xci/widgets/Icon.h>
-#include <xci/core/Vfs.h>
+#include <xci/vfs/Vfs.h>
 #include <xci/core/log.h>
 #include <xci/config.h>
