+++ tools/fire_script/Program.h
@@ -1,7 +1,7 @@
 // Program.h created on 2021-03-20 as part of xcikit project
 // https://github.com/rbrich/xcikit
 //
-// Copyright 2021 Radek Brich
+// Copyright 2021–2023 Radek Brich
 // Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #ifndef XCI_SCRIPT_TOOL_PROGRAM_H
@@ -12,7 +12,7 @@
 #include "ReplCommand.h"
 #include "Options.h"
 #include <xci/core/EditLine.h>
-#include <xci/core/Vfs.h>
+#include <xci/vfs/Vfs.h>
 
 #include <string_view>
